+++ svelte/demos/routes.js
@@ -1,5 +1,6 @@
 import BasicInit from "./cases/BasicInit.svelte";
 import MarkdownInit from "./cases/MarkdownInit.svelte";
+import CustomContent from "./cases/CustomContent.svelte";
 import Events from "./cases/Events.svelte";
 import Locale from "./cases/Locale.svelte";
 import BackendUrl from "./cases/BackendUrl.svelte";
@@ -9,6 +10,7 @@ import BackendResolve from "./cases/BackendResolve.svelte";
 export const links = [
 	["/base/:skin", "Comments basic", BasicInit, "BasicInit"],
 	["/markdown/:skin", "Markdown content", MarkdownInit, "MarkdownInit"],
+	["/custom-render/:skin", "Custom content", CustomContent, "CustomContent"],
 	["/events/:skin", "Events", Events, "Events"],
 	["/locale/:skin", "Locales", Locale, "Locale"],
 	["/backend-url/:skin", "Save to backend", BackendUrl, "BackendUrl"],