+++ svelte/src/components/Messages/MarkdownPlain.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { html } from "../../libs/lima.es.js";
-	export let text;
+	let { content } = $props();
 </script>
 
 <!-- eslint-disable-next-line svelte/no-at-html-tags -->
-{@html html(text)}
+{@html html(content)}