+++ svelte/src/index.js
@@ -1,5 +1,6 @@
 import Comments from "./components/Comments.svelte";
 import Willow from "./themes/Willow.svelte";
 import WillowDark from "./themes/WillowDark.svelte";
+import Material from "./themes/Material.svelte";
 
-export { Comments, Willow, WillowDark };
+export { Comments, Willow, WillowDark, Material };