+++ svelte/types/index.d.ts
@@ -1,0 +1,63 @@
+import type { Component } from "svelte";
+
+export interface IUser {
+	id: string | number;
+	name?: string;
+	avatar?: string;
+	color?: string;
+}
+
+export interface IComment {
+	id?: string | number;
+	content: string;
+	author?: IUser;
+	user?: string | number;
+	date?: Date;
+	format?: "text" | "markdown" | FormatComponent;
+}
+
+export interface IChange {
+	action: "add" | "update" | "delete";
+	id?: string | number;
+	comment?: IComment;
+	value: IComment[];
+	originalValue: IComment[] | string | number;
+}
+
+export type FormatComponent = Component<{
+	content: string;
+}>;
+
+export type RenderComponent = Component<{
+	owned?: string | number;
+	edit?: string | number;
+	author: IUser;
+	date: Date;
+}>;
+
+export declare const Comments: Component<{
+	ondata?: (value: string | number) => Promise<IComment[]> | IComment[];
+	onchange?: (ev: IChange) => void;
+	value?: IComment[] | string | number;
+	readonly?: boolean;
+	render?: "bubbles" | "flow" | RenderComponent;
+	format?: "text" | "markdown" | FormatComponent;
+	users?: IUser[];
+	activeUser?: string | number | IUser;
+	focus?: boolean;
+}>;
+
+export declare const Willow: Component<{
+	fonts?: boolean;
+	children?: () => any;
+}>;
+
+export declare const WillowDark: Component<{
+	fonts?: boolean;
+	children?: () => any;
+}>;
+
+export declare const Material: Component<{
+	fonts?: boolean;
+	children?: () => any;
+}>;